+++ exoptima/config/computation.py
@@ -1,13 +1,23 @@
 # computation defaults
 
-from astropy import units as u
+# default sp type from the drop_down
+DEFAULT_SP_TYPE = "G2"
+
+DEFAULT_EXPOSURE_TIME = 60.0
+
+# default instrument
+DEFAULT_INSTRUMENT = "EXOTICA"
 
 # default observing constraints
-
 DEFAULT_MAX_AIRMASS = 2.0
-DEFAULT_MIN_DURATION = 1 * u.hour
+DEFAULT_MIN_DURATION_HOUR = 1
 DEFAULT_MIN_MOON_SEP = 30.0
 DEFAULT_IGNORE_MOON_FLI = 0.05
+DEFAULT_SEEING_ARCSEC = 0.8
+DEFAULT_AIRMASS_RV = 1.2
+
+# default night limit defimition
+DEFAULT_NIGHT_DEFINITION = "Nautical twilight"
 
 # step in nights used in "Year" observability mode
 YEAR_OBS_NIGHTSTEP = 7