+++ exoptima/core/state.py
@@ -10,7 +10,7 @@ import numpy as np
 from astropy import units as u
 from astropy.time import Time
 
-from exoptima.config.instruments import Instrument
+from exoptima.config.instruments import Instrument, INSTRUMENTS
 
 # ----------------------------------------
 # Objects used to store params and results
@@ -40,12 +40,16 @@ class PlanetParameters:
 @dataclass
 class ObservingConditions:
     # airmass constraints
-    max_airmass: float = 2.0
-    min_duration: u.Quantity = 1 * u.hour
+    max_airmass: float
+    min_duration: u.Quantity
 
     # Moon-related constraints
-    min_moon_separation: float = 30.0      # degrees
-    ignore_moon_if_fli_above: float = 0.5  # fraction (0–1)
+    min_moon_separation: float           # degrees
+    ignore_moon_if_fli_above: float      # fraction (0–1)
+
+    # instantaneous seeing and airmass
+    seeing_arcsec: float
+    airmass_rv: float
 
 @dataclass
 class ObservabilityResult:
@@ -69,6 +73,18 @@ class NightObservability:
 @dataclass(frozen=True)
 class MultiNightObservability:
     nights: Sequence[NightObservability]
+
+# ----------------------------------------
+#   Valid entris for string inputs
+# ----------------------------------------
+
+VALID_SPTYPES = ["G2", "K2", "M2"]
+
+VALID_NIGHT_DEFINITIONS = [
+    "Sunset to sunrise",
+    "Nautical twilight",
+    "Astronomical twilight",
+]
 
 # ----------------------------------------
 #   Master State
@@ -178,11 +194,7 @@ class AppState(param.Parameterized):
     # ---------------------------------
     night_definition = param.ObjectSelector(
         default="Nautical twilight",
-        objects=[
-            "Sunset to sunrise",
-            "Nautical twilight",
-            "Astronomical twilight",
-        ],
+        objects=VALID_NIGHT_DEFINITIONS,
         doc="Definition of night boundaries for observability",
     )
 
@@ -201,15 +213,50 @@ class AppState(param.Parameterized):
 
         # Import here to avoid config ↔ state circular dependency
         from exoptima.config.computation import (
+            DEFAULT_SP_TYPE,
+            DEFAULT_EXPOSURE_TIME,
+            DEFAULT_INSTRUMENT,
             DEFAULT_MAX_AIRMASS,
-            DEFAULT_MIN_DURATION,
+            DEFAULT_MIN_DURATION_HOUR,
             DEFAULT_MIN_MOON_SEP,
             DEFAULT_IGNORE_MOON_FLI,
+            DEFAULT_SEEING_ARCSEC,
+            DEFAULT_AIRMASS_RV,
+            DEFAULT_NIGHT_DEFINITION,
         )
+
+        if DEFAULT_SP_TYPE in VALID_SPTYPES:
+            self.sp_type = DEFAULT_SP_TYPE
+        else:
+            print(
+                f"Warning: DEFAULT_SPTYPE='{DEFAULT_SP_TYPE}' is invalid. "
+                f"Falling back to 'G2'."
+            )
+            self.sp_type = "G2"
+
+        self.exposure_time=DEFAULT_EXPOSURE_TIME
+
+        if DEFAULT_INSTRUMENT not in INSTRUMENTS:
+            raise ValueError(
+                f"DEFAULT_INSTRUMENT='{DEFAULT_INSTRUMENT}' is not defined in INSTRUMENTS"
+            )
+
+        self.instrument=INSTRUMENTS[DEFAULT_INSTRUMENT]
 
         self.conditions = ObservingConditions(
             max_airmass=DEFAULT_MAX_AIRMASS,
-            min_duration=DEFAULT_MIN_DURATION,
+            min_duration=DEFAULT_MIN_DURATION_HOUR * u.hour,
             min_moon_separation=DEFAULT_MIN_MOON_SEP,
             ignore_moon_if_fli_above=DEFAULT_IGNORE_MOON_FLI,
+            seeing_arcsec=DEFAULT_SEEING_ARCSEC,
+            airmass_rv=DEFAULT_AIRMASS_RV,
         )
+
+        if DEFAULT_NIGHT_DEFINITION in VALID_NIGHT_DEFINITIONS:
+            self.night_definition = DEFAULT_NIGHT_DEFINITION
+        else:
+            print(
+                f"Warning: DEFAULT_NIGHT_DEFINITION='{DEFAULT_NIGHT_DEFINITION}' "
+                f"is invalid. Falling back to 'Nautical twilight'."
+            )
+            self.night_definition = "Nautical twilight"