+++ exoptima/tabs/interface.py
@@ -6,10 +6,11 @@ from exoptima.core.state import AppState
 from exoptima.config.layout import BUTTON_WIDTH, BUTTON_HEIGHT
 
 from exoptima.tabs.controls import (
-    make_star_tab, make_instrument_tab, make_observing_conditions_tab, make_planet_rv_tab, make_time_tab)
+    make_star_tab, make_instrument_tab, make_observing_conditions_tab,
+    make_planet_rv_tab, make_time_tab)
 from exoptima.tabs.display import (
-    make_daily_observability_tab, make_monthly_observability_tab, make_yearly_observability_tab,
-    make_precision_tab)
+    make_daily_observability_tab, make_monthly_observability_tab,
+    make_yearly_observability_tab, make_precision_tab)
 
 # ------------------------------------------------------------------
 # Header